Extract sharable prettier config from eslint config (PR #55)
This breaks out the GrowFlow prettier rules into a separate prettier-config package from the eslint-config package. Doing it this way as a separate package has some advantages over including the prettier config with the eslint config. The main one is if someone is using a prettier plugin (and not an eslint plugin) in their editor, they can still take advantage of the shared configuration.

This will avoid issues where someone could format a document with one set of rules in their editor, but running eslint --fix would format with a different set of rules.

This also includes previously-required peer/dev dependencies for the eslint config as regular dependencies. It makes it much easier to consume since a consuming app only has to install the single @growflow/eslint-config package.

I think we were previously copying what airbnb’s shared config did (requiring the sub deps as peers) but I think their use case is different than ours. For a public config that is highly consumed like theirs, I think that strategy makes sense. But for our config, where we are only using it internally, including the dependencies provides better consistency and makes it easier to provide upgrades across the whole org in the future.

I’ve updated the READMEs for both packages to reflect how to install and use it now.

Approved-by: Christian Saiki