🚸 tweak shared configs (eslint, typescript) (PR #95)
This continues to refine the shared eslint config based on usage in the Tenant backend and frontend.

In particular, these changes are meant to avoid anys from propagating throughout the codebase (but also remove annoying warnings everywhere).

It does this by making no-explicit-any and no-unsafe-return an error (these were previously warning) while also disabling explicit-module-boundary-types. The end result of this is that it should no longer show warnings/errors for implicitly typed items but it will show errors if you try to make something any (either explicitly or implicitly).

That last point is accomplished with a shared TypeScript config (since that is the only place we can enforce that) via noImplicitAny. growflow/tsconfig now exists and can be extended in projects that want to use it.