🔧 chore: babel be gone (#5)
This removes babel and replaces with the TS compiler directly in our jest and eslint packages. I won't bore you with a rehash of all the reasons here. See the associated retail PR and also this in the context of our jest config.

This removes our custom babel preset along with our base tsconfig (v7 will be the last version of those). We will be moving to using the TS compiler directly and possibly to swc later. We will also be making use of community-maintained TS configs.